+++ packages/lila-rollup-config/example/app/components/second/index.js
@@ -1,3 +1,3 @@
-import './index.scss';
-
 console.log('second');
+
+export default {};