+++ packages/lila-rollup-config/example/app/src/render.jsx
@@ -1,5 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import Hello from "./hello.jsx";
-
-ReactDOM.render(<Hello name="lila" />, document.getElementById('root'));