+++ packages/lila-rollup/example/app/lila.js
@@ -3,9 +3,7 @@ import rollupPlugin from '../../lib';
 export default lila => {
   const { setSetting } = lila;
 
-  setSetting('rollupConfigGenerator', () => options => {
-    const { cmd } = options;
-
+  setSetting('rollupConfigGenerator', () => () =>
     // console.log('rollupConfigGenerator/rollup', rollup);
     // console.log('rollupConfigGenerator/entry', options.entry);
     // console.log('rollupConfigGenerator/args', options.args);
@@ -14,18 +12,15 @@ export default lila => {
     // console.log('rollupConfigGenerator/config', options.config);
     // console.log('rollupConfigGenerator/lila', options.lila);
 
-    const isDev = cmd === 'start';
-    const dir = isDev ? 'dev' : 'lib';
-
-    return {
+    ({
       input: `${__dirname}/src/index.js`,
       output: {
-        file: `${__dirname}/${dir}/index.js`,
+        file: `${__dirname}/lib/index.js`,
         format: 'cjs',
         sourcemap: !0,
       },
-    };
-  });
+    })
+  );
 
   setSetting('getEntries', () =>
     // console.log(`getEntries/dir: ${dir}`);