+++ packages/lila-webpack-lib-config/example/app/demo/index.js
@@ -1,4 +1,5 @@
+/* eslint-disable import/no-unresolved */
 import 'bootstrap/dist/css/bootstrap.css';
 // import '../src';
 import '../lib/cjs';
-import '../lib/css.css';
+import '../lib/style.css';